refactor: clean up Toolbar
